MOV: Functions to upper level Replace fixed assignments by macros

+++ firmware/at32f415/app/inc/flash.h
@@ -29,8 +29,8 @@ typedef struct {
     const uint32_t opt;
     flash_res_t (*init)(void);
     flash_res_t (*erase)(uint32_t sector);
-    flash_res_t (*read)(uint8_t *data, uint32_t addr, uint16_t len);
-    flash_res_t (*write)(const uint8_t *data, uint32_t addr, uint16_t len);
+    flash_res_t (*read)(uint8_t *data, uint32_t addr, uint32_t len);
+    flash_res_t (*write)(const uint8_t *data, uint32_t addr, uint32_t len);
 }flash_t;
 
 #endif
